Rename interface and class name
